fixed #3

+++ src/main/java/BmiService.java
@@ -1,6 +1,6 @@
 public class BmiService {
     double calculate(double masskg, double growthm) {
-        double bmi = (double) (masskg / (growthm / 100 * growthm / 100));
+        int bmi = (int) (masskg / (growthm / 100 * growthm / 100));
         return bmi;
     }
 }
